refactor(cli) adapt pattern
